Split tests to follow other test convention - WinAttribution.test tests pure code, ClanSpectatorConversion.test properly tests core
